Classe Vertice feita #erro corrigido

+++ src/Vertice.cpp
@@ -13,6 +13,6 @@ void Vertice::calcularW_v (double v1, double v2, double v3, double v4) {
 	W_v = (G_v * v1) + (D_v * v2) + (M_v * v3) + (P_v * v4);
 }
 
-int Vertice::getW_v () {
+double Vertice::getW_v () {
 	return W_v;
 }
